+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/journeys/landlordRegistration/tasks/LandlordRegistrationTask.kt
@@ -1,29 +1,41 @@
 package uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.tasks
 
 import uk.gov.communities.prsdb.webapp.annotations.webAnnotations.JourneyFrameworkComponent
+import uk.gov.communities.prsdb.webapp.config.managers.FeatureFlagManager
+import uk.gov.communities.prsdb.webapp.constants.ORGANISATION_LANDLORD_REGISTRATION
 import uk.gov.communities.prsdb.webapp.journeys.Destination
+import uk.gov.communities.prsdb.webapp.journeys.OrParents
 import uk.gov.communities.prsdb.webapp.journeys.StepLifecycleOrchestrator
 import uk.gov.communities.prsdb.webapp.journeys.Task
 import uk.gov.communities.prsdb.webapp.journeys.builders.JourneyBuilder.Companion.journey
+import uk.gov.communities.prsdb.webapp.journeys.builders.SubJourneyBuilder
 import uk.gov.communities.prsdb.webapp.journeys.hasOutcome
 import uk.gov.communities.prsdb.webapp.journeys.isComplete
 import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.states.LandlordRegistrationState
-import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.stepConfig.CountryOfResidenceMode
 import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.stepConfig.CountryOfResidenceStep
 import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.stepConfig.DateOfBirthStep
 import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.stepConfig.EmailStep
-import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.stepConfig.NonEnglandOrWalesAddressStep
+import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.stepConfig.LandlordTypeMode
+import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.stepConfig.LandlordTypeStep
 import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.stepConfig.PhoneNumberStep
 import uk.gov.communities.prsdb.webapp.journeys.landlordRegistration.stepConfig.PrivacyNoticeStep
 import uk.gov.communities.prsdb.webapp.journeys.shared.states.CheckYourAnswersJourneyState.Companion.checkAnswerStep
 import uk.gov.communities.prsdb.webapp.journeys.shared.states.CheckYourAnswersJourneyState.Companion.checkAnswerTask
-import uk.gov.communities.prsdb.webapp.journeys.shared.stepConfig.AbstractCheckYourAnswersStep
 import uk.gov.communities.prsdb.webapp.journeys.shared.stepConfig.LookupAddressStep
 import uk.gov.communities.prsdb.webapp.journeys.shared.stepConfig.NameStep
 
 @JourneyFrameworkComponent
-class LandlordRegistrationTask : Task<LandlordRegistrationState>() {
+class LandlordRegistrationTask(
+    private val featureFlagManager: FeatureFlagManager,
+) : Task<LandlordRegistrationState>() {
     override fun makeSubJourney(state: LandlordRegistrationState) =
+        if (featureFlagManager.checkFeature(ORGANISATION_LANDLORD_REGISTRATION)) {
+            makeOrgLandlordSubJourney(state)
+        } else {
+            makeIndividualLandlordSubJourney(state)
+        }
+
+    private fun makeIndividualLandlordSubJourney(state: LandlordRegistrationState): SubJourneyBuilder<LandlordRegistrationState> =
         subJourney(state) {
             step(journey.privacyNoticeStep) {
                 routeSegment(PrivacyNoticeStep.ROUTE_SEGMENT)
@@ -31,44 +43,52 @@ class LandlordRegistrationTask : Task<LandlordRegistrationState>() {
             }
             task(journey.identityTask) {
                 parents { journey.privacyNoticeStep.isComplete() }
-                nextStep { journey.emailStep }
+                nextStep { journey.individualLandlordRegistrationTask.firstStep }
             }
-            step(journey.emailStep) {
-                routeSegment(EmailStep.ROUTE_SEGMENT)
+            task(journey.individualLandlordRegistrationTask) {
                 parents { journey.identityTask.isComplete() }
-                nextStep { journey.phoneNumberStep }
-            }
-            step(journey.phoneNumberStep) {
-                routeSegment(PhoneNumberStep.ROUTE_SEGMENT)
-                parents { journey.emailStep.isComplete() }
-                nextStep { journey.countryOfResidenceStep }
-            }
-            step(journey.countryOfResidenceStep) {
-                routeSegment(CountryOfResidenceStep.ROUTE_SEGMENT)
-                parents { journey.phoneNumberStep.isComplete() }
-                nextStep { mode ->
-                    when (mode) {
-                        CountryOfResidenceMode.ENGLAND_OR_WALES -> journey.addressTask.firstStep
-                        CountryOfResidenceMode.NON_ENGLAND_OR_WALES -> journey.nonEnglandOrWalesAddressStep
-                    }
-                }
-            }
-            step(journey.nonEnglandOrWalesAddressStep) {
-                routeSegment(NonEnglandOrWalesAddressStep.ROUTE_SEGMENT)
-                parents { journey.countryOfResidenceStep.hasOutcome(CountryOfResidenceMode.NON_ENGLAND_OR_WALES) }
-                noNextDestination()
-            }
-            task(journey.addressTask) {
-                parents { journey.countryOfResidenceStep.hasOutcome(CountryOfResidenceMode.ENGLAND_OR_WALES) }
-                nextStep { journey.cyaStep }
-            }
-            step(journey.cyaStep) {
-                routeSegment(AbstractCheckYourAnswersStep.ROUTE_SEGMENT)
-                parents { journey.addressTask.isComplete() }
                 nextStep { exitStep }
             }
             exitStep {
-                parents { journey.cyaStep.isComplete() }
+                parents { journey.individualLandlordRegistrationTask.isComplete() }
+            }
+        }
+
+    private fun makeOrgLandlordSubJourney(state: LandlordRegistrationState): SubJourneyBuilder<LandlordRegistrationState> =
+        subJourney(state) {
+            step(journey.privacyNoticeStep) {
+                routeSegment(PrivacyNoticeStep.ROUTE_SEGMENT)
+                nextStep { journey.identityTask.firstStep }
+            }
+            task(journey.identityTask) {
+                parents { journey.privacyNoticeStep.isComplete() }
+                nextStep { journey.landlordTypeStep }
+            }
+            step(journey.landlordTypeStep) {
+                routeSegment(LandlordTypeStep.ROUTE_SEGMENT)
+                parents { journey.identityTask.isComplete() }
+                nextStep { mode ->
+                    when (mode) {
+                        LandlordTypeMode.INDIVIDUAL -> journey.individualLandlordRegistrationTask.firstStep
+                        LandlordTypeMode.ORGANISATION -> journey.orgLandlordRegistrationTask.firstStep
+                    }
+                }
+            }
+            task(journey.orgLandlordRegistrationTask) {
+                parents { journey.landlordTypeStep.hasOutcome(LandlordTypeMode.ORGANISATION) }
+                nextStep { exitStep }
+            }
+            task(journey.individualLandlordRegistrationTask) {
+                parents { journey.landlordTypeStep.hasOutcome(LandlordTypeMode.INDIVIDUAL) }
+                nextStep { exitStep }
+            }
+            exitStep {
+                parents {
+                    OrParents(
+                        journey.individualLandlordRegistrationTask.isComplete(),
+                        journey.orgLandlordRegistrationTask.isComplete(),
+                    )
+                }
             }
         }
 
@@ -84,16 +104,32 @@ class LandlordRegistrationTask : Task<LandlordRegistrationState>() {
                 }
                 configureFirst { backDestination { journey.returnToCyaPageDestination } }
                 when (checkingAnswersFor) {
-                    NameStep.ROUTE_SEGMENT -> checkAnswerStep(journey.nameStep, NameStep.ROUTE_SEGMENT)
-                    DateOfBirthStep.ROUTE_SEGMENT -> checkAnswerStep(journey.dateOfBirthStep, DateOfBirthStep.ROUTE_SEGMENT)
-                    EmailStep.ROUTE_SEGMENT -> checkAnswerStep(journey.emailStep, EmailStep.ROUTE_SEGMENT)
-                    PhoneNumberStep.ROUTE_SEGMENT -> checkAnswerStep(journey.phoneNumberStep, PhoneNumberStep.ROUTE_SEGMENT)
-                    CountryOfResidenceStep.ROUTE_SEGMENT ->
+                    NameStep.ROUTE_SEGMENT -> {
+                        checkAnswerStep(journey.nameStep, NameStep.ROUTE_SEGMENT)
+                    }
+
+                    DateOfBirthStep.ROUTE_SEGMENT -> {
+                        checkAnswerStep(journey.dateOfBirthStep, DateOfBirthStep.ROUTE_SEGMENT)
+                    }
+
+                    EmailStep.ROUTE_SEGMENT -> {
+                        checkAnswerStep(journey.emailStep, EmailStep.ROUTE_SEGMENT)
+                    }
+
+                    PhoneNumberStep.ROUTE_SEGMENT -> {
+                        checkAnswerStep(journey.phoneNumberStep, PhoneNumberStep.ROUTE_SEGMENT)
+                    }
+
+                    CountryOfResidenceStep.ROUTE_SEGMENT -> {
                         checkAnswerStep(
                             journey.countryOfResidenceStep,
                             CountryOfResidenceStep.ROUTE_SEGMENT,
                         )
-                    LookupAddressStep.ROUTE_SEGMENT -> checkAnswerTask(journey.addressTask)
+                    }
+
+                    LookupAddressStep.ROUTE_SEGMENT -> {
+                        checkAnswerTask(journey.addressTask)
+                    }
                 }
                 step(journey.finishCyaStep) {
                     initialStep()
