--- src/main/kotlin/uk/gov/communities/prsdb/webapp/journeys/organisationLandlordRegistration/steps/OrgTypeStepConfig.kt
@@ -1,29 +1,0 @@
-package uk.gov.communities.prsdb.webapp.journeys.organisationLandlordRegistration.steps
-
-import uk.gov.communities.prsdb.webapp.annotations.webAnnotations.JourneyFrameworkComponent
-import uk.gov.communities.prsdb.webapp.journeys.AbstractRequestableStepConfig
-import uk.gov.communities.prsdb.webapp.journeys.JourneyState
-import uk.gov.communities.prsdb.webapp.journeys.JourneyStep.RequestableStep
-import uk.gov.communities.prsdb.webapp.journeys.shared.Complete
-import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.NoInputFormModel
-
-@JourneyFrameworkComponent
-class OrgTypeStepConfig : AbstractRequestableStepConfig<Complete, NoInputFormModel, JourneyState>() {
-    override val formModelClass = NoInputFormModel::class
-
-    override fun getStepSpecificContent(state: JourneyState) =
-        mapOf("todoComment" to "TODO: PDJB-1137 - What is the type of the organisation")
-
-    override fun chooseTemplate(state: JourneyState) = "forms/todo"
-
-    override fun mode(state: JourneyState) = getFormModelFromStateOrNull(state)?.let { Complete.COMPLETE }
-}
-
-@JourneyFrameworkComponent
-final class OrgTypeStep(
-    stepConfig: OrgTypeStepConfig,
-) : RequestableStep<Complete, NoInputFormModel, JourneyState>(stepConfig) {
-    companion object {
-        const val ROUTE_SEGMENT = "organisation-type"
-    }
-}
