+++ src/main/resources/messages/registerAsALandlord.yml
@@ -1,5 +1,6 @@
 title: Register as a Landlord
 heading: 'Register your rental property: private beta'
+caption: Register as a landlord
 intro:
   one: Landlords will soon need to use GOV.UK to register their rental properties. This will help local councils understand the rental properties in their area.
   two: "You’ve been invited to test how registering properties will work. You can share feedback at any point. We’ll use your feedback to improve the service."
@@ -70,7 +71,6 @@ privacyNotice:
     label: I confirm I have read the privacy notice for this private beta
   oneLoginInset: "When you continue, you\u2019ll be asked to prove your identity using your GOV.UK One Login."
 landlordType:
-  caption: Register as a landlord
   fieldSetHeading: Are you registering as an individual or an organisation?
   partnership:
     summary: Registering a partnership
